refactor: remove filterHasFocus data from store
