+++ src/store/selectors/SessionSelectors.ts
@@ -11,8 +11,6 @@ import { UserType } from '../../types/store/UserTypes';
 export const getSelectedConversation = (store: RootStore, roomId: string): boolean =>
 	store.session.selectedRoomOneToOneGroup === roomId;
 
-export const getSidebarFilterHasFocus = (store: RootStore): boolean => store.session.filterHasFocus;
-
 export const getAttribute = (
 	store: RootStore,
 	attributeName: keyof AttributesList
