+++ src/store/slices/SessionStoreSlice.ts
@@ -10,8 +10,8 @@ import { produce } from 'immer';
 import { StateCreator } from 'zustand';
 
 import ChatExporter from '../../settings/components/chatExporter/ChatExporter';
-import { AttributesList, ExportStatus } from '../../types/store/SessionTypes';
-import { RootStore, SessionStoreSlice } from '../../types/store/StoreTypes';
+import { AttributesList, ExportStatus, SessionStoreSlice } from '../../types/store/SessionTypes';
+import { RootStore } from '../../types/store/StoreTypes';
 import { UserType } from '../../types/store/UserTypes';
 import UserDataRetriever from '../../utils/UserDataRetriever';
 
@@ -21,9 +21,7 @@ export const useSessionStoreSlice: StateCreator<
 	[],
 	SessionStoreSlice
 > = (set) => ({
-	session: {
-		filterHasFocus: false
-	},
+	session: {},
 	setLoginInfo: (id: string, name: string, displayName?: string, userType?: UserType): void => {
 		set(
 			produce((draft: RootStore) => {
@@ -37,8 +35,7 @@ export const useSessionStoreSlice: StateCreator<
 						chats_be: undefined,
 						xmpp: undefined,
 						websocket: undefined
-					},
-					filterHasFocus: draft.session.filterHasFocus
+					}
 				};
 				UserDataRetriever.getDebouncedUser(id, true);
 			}),
@@ -94,15 +91,6 @@ export const useSessionStoreSlice: StateCreator<
 			}),
 			false,
 			'SESSION/SET_SELECTED_ROOM_ONE_TO_ONE_GROUP'
-		);
-	},
-	setFilterHasFocus: (hasFocus: boolean): void => {
-		set(
-			produce((draft: RootStore) => {
-				draft.session.filterHasFocus = hasFocus;
-			}),
-			false,
-			'SESSION/SET_FILTER_FOCUS'
 		);
 	},
 	setCustomLogo: (logo: string | false): void => {
