+++ src/types/store/StoreTypes.ts
@@ -4,8 +4,6 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import { AccountSettingsAttrs } from '@zextras/carbonio-shell-ui/lib/types/account';
-
 import { ActiveConversationsMap, FileToUpload, messageActionType } from './ActiveConversationTypes';
 import {
 	ActiveMeetingMap,
@@ -30,9 +28,9 @@ import {
 	TextMessage
 } from './MessageTypes';
 import { RoomsMap } from './RoomTypes';
-import { ExportStatus, Session } from './SessionTypes';
+import { SessionStoreSlice } from './SessionTypes';
 import { UnreadsMap } from './UnreadsCounterTypes';
-import { UsersMap, UserType } from './UserTypes';
+import { UsersMap } from './UserTypes';
 import { MeetingBe } from '../network/models/meetingBeTypes';
 import { MemberBe, RoomBe } from '../network/models/roomBeTypes';
 import { UserBe } from '../network/models/userBeTypes';
@@ -85,18 +83,6 @@ export type MessagesStoreSlice = {
 	) => void;
 	setPlaceholderMessage: (fields: PlaceholderFields) => void;
 	removePlaceholderMessage: (roomId: string, messageId: string) => void;
-};
-
-export type SessionStoreSlice = {
-	session: Session;
-	setFilterHasFocus: (hasFocus: boolean) => void;
-	setLoginInfo: (id: string, name: string, displayName?: string, userType?: UserType) => void;
-	setSessionId: (sessionId: string) => void;
-	setAttributes: (attrs: AccountSettingsAttrs) => void;
-	setSelectedRoomOneToOneGroup: (id: string) => void;
-	setCustomLogo: (logo: string | false) => void;
-	setChatExporting: (roomId?: string) => void;
-	setChatExportStatus: (status: ExportStatus) => void;
 };
 
 export type MarkersStoreSlice = {
